social/service/requirements.txt
django==5.1.1
djangorestframework==3.15.2
psycopg==3.2.2
